event log limit not considering preexisting logs
Turns out the code that checks for event log size limitations forgot
to take in to account the total size of logs that already existed.  So
it would let you create the max size of new logs rather then new+old.

I fixed this by initializing the variable when the process starts.  I
also added a test case that catches the problem.